src/mvt/requirements.txt
torch
mistralai
faiss-cpu
requests
python-dotenv
langchain
langchain-community
langchain-mistralai
beautifulsoup4
transformers
fastapi
uvicorn
streamlit
streamlit-authenticator
youtube-transcript-api
pypdf
chardet
sentence-transformers